Constify tmpstr, refstr and result
-Use ssize_t as datatype argcount instead of int.

(cherry picked from commit 53c1b3fe26319b3eb914061665acb2c1b8e4bb02)

+++ lib/dvb/epgcache.cpp
@@ -1624,7 +1624,7 @@ int handleEvent(eServiceEvent *ptr, ePyObject dest_list, const char* argstring, 
 PyObject *eEPGCache::lookupEvent(ePyObject list, ePyObject convertFunc)
 {
 	ePyObject convertFuncArgs;
-	int argcount=0;
+	ssize_t argcount=0;
 	const char *argstring=NULL;
 	if (!PyList_Check(list))
 	{
